api/pyproject.toml
[project]
name = "farmaeasy-api"
version = "0.1.0"
description = "Sistema de Gestao de Farmacia - API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Robert", email = "robert@example.com" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "httpx>=0.28.0",
    "starlette-compress>=1.0.1",
    "cassandra-driver>=3.29.0",
    "redis>=5.2.0",
    # Authentication
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    "cryptography>=46.0.3",
    # File Upload & Storage
    "firebase-admin>=6.5.0",
    "python-multipart>=0.0.9",
    # Email (Gmail API)
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "mutmut>=3.2.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "asgi-lifespan>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "mutmut>=3.2.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "asgi-lifespan>=2.1.0",
]

# ==============================================================================
# RUFF CONFIGURATION
# ==============================================================================
[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "htmlcov",
    "dist",
    "build",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "TID",    # flake8-tidy-imports
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PYI",    # flake8-pyi
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    "INT",    # flake8-gettext
    "FLY",    # flynt
    "PERF",   # Perflint
    "FURB",   # refurb
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "S101",    # assert usage (ok in tests)
    "RUF012",  # mutable class attributes
]

[tool.ruff.lint.per-file-ignores]
# Tests: asserts, unused args, magic values, prints, hardcoded secrets, imports inside funcs
"tests/**/*.py" = ["S101", "ARG", "PLR2004", "T20", "S105", "S106", "PLC0415", "ERA001", "SIM117"]
"tests/conftest.py" = ["ARG", "S105", "S106", "PLC0415", "ERA001", "SIM117"]
# Circular import prevention - imports inside functions are intentional
"src/core/middleware.py" = ["PLC0415", "PLR2004"]
"src/core/logging.py" = ["ARG001", "PLC0415"]  # structlog processor signature requires these args
"src/main.py" = ["PLC0415", "ARG001", "ERA001", "PERF401"]  # exception handlers require request arg
"src/health/router.py" = ["PLC0415"]
"src/core/database/cassandra.py" = ["ERA001"]  # Informational comments are not commented-out code
# Settings defaults with hardcoded values are expected
"src/config/settings.py" = ["S104", "S105", "S106"]
# Auth module - OAuth "Bearer" token_type and CQL statements are not secrets/injections
"src/auth/models.py" = ["S105", "S608"]  # CQL table creation strings
"src/auth/schemas.py" = ["S105"]  # OAuth2 "Bearer" token_type default
"src/auth/service.py" = ["S106", "S608"]  # OAuth2 "Bearer" token_type parameter + CQL queries
# Courses module - CQL statements and global service getters pattern
"src/courses/models.py" = ["S608"]  # CQL table creation strings
"src/courses/service.py" = ["S608", "RET505"]  # CQL prepared statements
"src/courses/dependencies.py" = ["PLW0603"]  # Global service getter pattern (same as auth)
# Comments module - CQL statements
"src/comments/models.py" = ["S608"]  # CQL table creation strings
"src/comments/service.py" = ["S608"]  # CQL prepared statements
# Progress module - CQL statements
"src/progress/models.py" = ["S608"]  # CQL table creation strings
"src/progress/service.py" = ["S608"]  # CQL prepared statements
# CLI scripts - print statements and magic values are expected
"scripts/**/*.py" = ["T20", "PLR2004", "PLR0912", "S105", "S106", "BLE001"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
docstring-quotes = "double"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

# ==============================================================================
# PYTEST CONFIGURATION
# ==============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--import-mode=importlib",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=75",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ==============================================================================
# COVERAGE CONFIGURATION
# ==============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "class .*\\bProtocol\\):",
    "\\.\\.\\.",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# ==============================================================================
# MUTMUT CONFIGURATION
# ==============================================================================
[tool.mutmut]
paths_to_mutate = "src/"
tests_dir = "tests/"
runner = "python -m pytest -x -q --tb=no"
